% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_sets.R
\name{default_parsers}
\alias{default_parsers}
\title{Default set of parsers}
\usage{
default_parsers()
}
\value{
List of trial types and respective parser functions
}
\description{
This list of used to automatically call respective parsers.
It is primarily used for auto-detection in \code{process_records}.
In future, it could be possible to hold alternative parsers for different
versions of jsPsych (or jspsychread).
}
\examples{
default_parsers()
# extending parsers
my_parsers <- default_parsers()
my_form_parser <- function(x) { x } # not useful but check parsers.R for inspiration
my_parsers[["survey_html_form"]] <- my_form_parser
}
