% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{demo_file}
\alias{demo_file}
\title{Demo files}
\usage{
demo_file(fn)
}
\arguments{
\item{fn}{File to read from the demo folder}
}
\value{
File path for given demo file

The list of available demo files:
\itemize{
\item jspsych-animation.json
\item jspsych-audio-button-response.json
\item jspsych-audio-keyboard-response.json
\item jspsych-audio-slider-response.json
\item jspsych-call-function.json
\item jspsych-canvas-button-response.json
\item jspsych-canvas-keyboard-response.json
\item jspsych-canvas-slider-response.json
\item jspsych-free-sort.json
\item jspsych-fullscreen.json
\item jspsych-html-button-response.json
\item jspsych-html-keyboard-response.json
\item jspsych-html-slider-response.json
\item jspsych-image-button-response.json
\item jspsych-image-keyboard-response.json
\item jspsych-image-slider-response.json
\item jspsych-instructions.json
\item jspsych-serial-reaction-time-mouse.json
\item jspsych-serial-reaction-time.json
\item jspsych-survey-html-form.json
\item jspsych-survey-likert.json
\item jspsych-survey-multi-choice.json
\item jspsych-survey-multi-select.json
\item jspsych-survey-text.json
\item jspsych-video-button-response.json
\item jspsych-video-keyboard-response.json
\item jspsych-video-slider-response.json
}
}
\description{
Returns a path for example data.
}
\examples{
demo_file("jspsych-html-button-response.json")
}
