% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jspsychread-package.R
\docType{package}
\name{jspsychread-package}
\alias{jspsychread}
\alias{jspsychread-package}
\title{jspsychread: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jirilukavsky/jspsychread}
  \item Report bugs at \url{https://github.com/jirilukavsky/jspsychread/issues}
}

}
\author{
\strong{Maintainer}: Jiri Lukavsky \email{lukavsky@praha.psu.cas.cz} (\href{https://orcid.org/0000-0002-1082-229X}{ORCID})

}
\keyword{internal}
