% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list2_to_matrix}
\alias{list2_to_matrix}
\title{Convert list of lists into a matrix}
\usage{
list2_to_matrix(x, byrow = T)
}
\arguments{
\item{x}{List of lists, expecting all inner-lists of the same length}

\item{byrow}{logical. If TRUE (the default) the matrix is filled by rows,
otherwise the matrix is filled by columns.}
}
\value{
Matrix
}
\description{
We expect the same representation as in serial-reaction-time-mouse:
Each inner array represents a single row.
The entries in the inner arrays represent the columns.
}
\examples{
l <- list(list(1:3), list(4:6))
list2_to_matrix(l)
}
