% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_column_of_vectors}
\alias{list_column_of_vectors}
\title{Wrapper to create a list column with vectors in parser results}
\usage{
list_column_of_vectors(x, replace_null = T)
}
\arguments{
\item{x}{Vector}

\item{replace_null}{True if NULL should be replaced with NA}
}
\value{
List containing x
}
\description{
Mainly used for documenting the parser code.
}
\examples{
library(tibble)
tb <- tibble(i = 1, j = list_column_of_vectors(1:10))
tb
tb$j[[1]]
}
