% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nested_tibble}
\alias{nested_tibble}
\title{Wrapper to nest a tibble into parser result}
\usage{
nested_tibble(x)
}
\arguments{
\item{x}{Tibble}
}
\value{
List containing x
}
\description{
Mainly used for documenting the parser code.
}
\examples{
library(tibble)
ta <- tibble(a = 1:2)
tb <- tibble(i = 1, j = nested_tibble(ta))
ta
tb
}
