% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nested_tibble_from_list}
\alias{nested_tibble_from_list}
\title{Wrapper to nest a tibble from a list into parser result}
\usage{
nested_tibble_from_list(x)
}
\arguments{
\item{x}{List which can be transformed to tibble using \code{bind_rows}}
}
\value{
List containing tibble created from x
}
\description{
Mainly used for documenting the parser code.
}
\examples{
library(tibble)
d <- list(list("a" = 1, "b" = "apple"), list("a" = 2, "b" = "banana"))
d
tb <- tibble(i = 1, j = nested_tibble_from_list(d))
tb
}
