% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_animation}
\alias{parse_animation}
\title{animation parser}
\usage{
parse_animation(d)
}
\arguments{
\item{d}{List with unprocessed trial data}
}
\value{
Single row tibble with results. Check \href{https://www.jspsych.org/7.3/plugins/animation}{jsPsych documentation}
for the list of available variables.
}
\description{
https://www.jspsych.org/7.3/plugins/animation
}
\examples{
\dontrun{
fn <- demo_file("jspsych-animation.json")
d  <- read_jspsych(fn)
anim_data <-
  d \%>\%
  filter(trial_type == trial_types$animation) \%>\%
  select(record, trial_index, raw) \%>\%
  process_records(.using = parse_animation) \%>\%
  unnest(processed)
anim_data \%>\% unnest(response)
anim_data \%>\% unnest(animation_sequence)
}
}
