% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_call_function}
\alias{parse_call_function}
\title{jspsych-call-function parser}
\usage{
parse_call_function(d)
}
\arguments{
\item{d}{List with unprocessed trial data

value    any    The return value of the called function.}
}
\value{
Single row tibble with results
}
\description{
https://www.jspsych.org/plugins/jspsych-call-function/
}
\examples{
fn <- system.file("testdata", "jspsych-call-function.json", package = "jspsychread")
d  <- read_jspsych(fn)
d$raw \%>\%
  purrr::map(~ parse_call_function(.x)) \%>\%
  dplyr::bind_rows()
}
