% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_categorize_animation}
\alias{parse_categorize_animation}
\title{categorize-animation parser}
\usage{
parse_categorize_animation(d)
}
\arguments{
\item{d}{List with unprocessed trial data}
}
\value{
Single row tibble with results. Check \href{https://www.jspsych.org/7.3/plugins/categorize-animation}{jsPsych documentation}
for the list of available variables.
}
\description{
https://www.jspsych.org/7.3/plugins/categorize-animation
}
\examples{
\dontrun{
fn <- demo_file("jspsych-categorize-animation.json")
d  <- read_jspsych(fn)
d \%>\%
  filter(trial_type == trial_types$categorize_animation) \%>\%
  select(record, trial_index, raw) \%>\%
  process_records(.using = parse_categorize_animation) \%>\%
  unnest(processed)
}
}
