% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_cloze}
\alias{parse_cloze}
\title{cloze parser}
\usage{
parse_cloze(d)
}
\arguments{
\item{d}{List with unprocessed trial data}
}
\value{
Single row tibble with results. Check \href{https://www.jspsych.org/7.3/plugins/cloze}{jsPsych documentation}
for the list of available variables.
}
\description{
https://www.jspsych.org/7.3/plugins/cloze
}
\examples{
\dontrun{
fn <- demo_file("jspsych-cloze.json")
d  <- read_jspsych(fn)
trials <-
  d \%>\%
  filter(trial_type == trial_types$cloze) \%>\%
  select(record, trial_index, raw) \%>\%
  process_records(.using = parse_cloze) \%>\%
  unnest(processed)
# for individual values
trials \%>\%
  unnest(response) \%>\%
  mutate(cell_value = flatten_chr(response))
}
}
