% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_free_sort}
\alias{parse_free_sort}
\title{free-sort parser}
\usage{
parse_free_sort(d)
}
\arguments{
\item{d}{List with unprocessed trial data}
}
\value{
Single row tibble with results. Check \href{https://www.jspsych.org/7.3/plugins/free-sort}{jsPsych documentation}
for the list of available variables.
}
\description{
https://www.jspsych.org/7.3/plugins/free-sort
}
\examples{
\dontrun{
fn <- demo_file("jspsych-free-sort.json")
d  <- read_jspsych(fn)
trials <-
  d \%>\%
  filter(trial_type == trial_types$free_sort) \%>\%
  select(record, trial_index, raw) \%>\%
  process_records(.using = parse_free_sort) \%>\%
  unnest(processed)
# for locations
trials \%>\% select(-raw, -moves) \%>\% unnest(locations)
# for moves
trials \%>\% select(-raw, -locations) \%>\% unnest(moves)
}
}
