% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_fullscreen}
\alias{parse_fullscreen}
\title{fullscreen parser}
\usage{
parse_fullscreen(d)
}
\arguments{
\item{d}{List with unprocessed trial data}
}
\value{
Single row tibble with results. Check \href{https://www.jspsych.org/7.3/plugins/fullscreen}{jsPsych documentation}
for the list of available variables.
}
\description{
https://www.jspsych.org/7.3/plugins/fullscreen
}
\examples{
\dontrun{
fn <- demo_file("jspsych-fullscreen.json")
d  <- read_jspsych(fn)
d \%>\%
  filter(trial_type == trial_types$fullscreen) \%>\%
  select(record, trial_index, raw) \%>\%
  process_records(.using = parse_fullscreen) \%>\%
  unnest(processed)
}
}
