% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_image_keyboard_response}
\alias{parse_image_keyboard_response}
\title{jspsych-image-keyboard-response parser}
\usage{
parse_image_keyboard_response(d)
}
\arguments{
\item{d}{List with unprocessed trial data

response    (numeric) -    Indicates which key the subject pressed.
rt    (numeric) -    The response time in milliseconds for the subject to make a response. The time is measured from when the stimulus first appears on the screen until the subject's response.
stimulus    (string) - The path of the image that was displayed.}
}
\value{
Single row tibble with results
}
\description{
https://www.jspsych.org/plugins/jspsych-image-keyboard-response/
}
