% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_maxdiff}
\alias{parse_maxdiff}
\title{maxdiff parser}
\usage{
parse_maxdiff(d)
}
\arguments{
\item{d}{List with unprocessed trial data}
}
\value{
Single row tibble with results. Check \href{https://www.jspsych.org/7.3/plugins/maxdiff}{jsPsych documentation}
for the list of available variables.
}
\description{
https://www.jspsych.org/7.3/plugins/maxdiff
}
\examples{
\dontrun{
fn <- demo_file("jspsych-maxdiff.json")
d  <- read_jspsych(fn)
d \%>\%
  filter(trial_type == trial_types$maxdiff) \%>\%
  select(record, trial_index, raw) \%>\%
  process_records(.using = parse_maxdiff) \%>\%
  unnest(processed)
}
}
