% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_preload}
\alias{parse_preload}
\title{preload parser}
\usage{
parse_preload(d)
}
\arguments{
\item{d}{List with unprocessed trial data}
}
\value{
Single row tibble with results. Check \href{https://www.jspsych.org/7.3/plugins/preload}{jsPsych documentation}
for the list of available variables.
}
\description{
https://www.jspsych.org/7.3/plugins/preload
}
\examples{
\dontrun{
fn <- demo_file("jspsych-preload.json")
d  <- read_jspsych(fn)
d \%>\%
  filter(trial_type == trial_types$preload) \%>\%
  select(record, trial_index, raw) \%>\%
  process_records(.using = parse_preload) \%>\%
  unnest(processed)
}
}
