% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_reconstruction}
\alias{parse_reconstruction}
\title{reconstruction parser}
\usage{
parse_reconstruction(d)
}
\arguments{
\item{d}{List with unprocessed trial data}
}
\value{
Single row tibble with results. Check \href{https://www.jspsych.org/7.3/plugins/reconstruction}{jsPsych documentation}
for the list of available variables.
}
\description{
https://www.jspsych.org/7.3/plugins/reconstruction
}
\examples{
\dontrun{
fn <- demo_file("jspsych-reconstruction.json")
d  <- read_jspsych(fn)
d \%>\%
  filter(trial_type == trial_types$reconstruction) \%>\%
  select(record, trial_index, raw) \%>\%
  process_records(.using = parse_reconstruction) \%>\%
  unnest(processed)
}
}
