% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_resize}
\alias{parse_resize}
\title{resize parser}
\usage{
parse_resize(d)
}
\arguments{
\item{d}{List with unprocessed trial data}
}
\value{
Single row tibble with results. Check \href{https://www.jspsych.org/7.3/plugins/resize}{jsPsych documentation}
for the list of available variables.
}
\description{
https://www.jspsych.org/7.3/plugins/resize
}
\examples{
\dontrun{
fn <- demo_file("jspsych-resize.json")
d  <- read_jspsych(fn)
d \%>\%
  filter(trial_type == trial_types$resize) \%>\%
  select(record, trial_index, raw) \%>\%
  process_records(.using = parse_resize) \%>\%
  unnest(processed)
}
}
