% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_serial_reaction_time_mouse}
\alias{parse_serial_reaction_time_mouse}
\title{serial-reaction-time-mouse parser}
\usage{
parse_serial_reaction_time_mouse(d)
}
\arguments{
\item{d}{List with unprocessed trial data}
}
\value{
Single row tibble with results. Check \href{https://www.jspsych.org/7.3/plugins/serial-reaction-time-mouse}{jsPsych documentation}
for the list of available variables.
}
\description{
https://www.jspsych.org/7.3/plugins/serial-reaction-time-mouse
}
\examples{
\dontrun{
fn <- demo_file("jspsych-serial-reaction-time-mouse.json")
d  <- read_jspsych(fn)
d \%>\%
  filter(trial_type == trial_types$serial_reaction_time_mouse) \%>\%
  select(record, trial_index, raw) \%>\%
  process_records(.using = parse_serial_reaction_time_mouse) \%>\%
  unnest(processed)
trials \%>\%
  select(-raw) \%>\%
  mutate(ncol = map_int(.$grid, ~ ncol(.)),
         nrow = map_int(.$grid, ~ nrow(.)))
}
}
