% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_single_record.R
\name{parse_single_record}
\alias{parse_single_record}
\title{Export all generic data from single session}
\usage{
parse_single_record(jsondata)
}
\arguments{
\item{jsondata}{Data in JSON format

Currently
\itemize{
\item \code{trial_type}    (string)    - The name of the plugin used to run the trial.
\item \code{trial_index} (numeric) - The index of the current trial across the whole experiment.
\item \code{time_elapsed}    (numeric)    - The number of milliseconds between the start of the experiment and when the trial ended.
\item \code{internal_node_id}    (string) - A string identifier for the current TimelineNode.
\item \code{raw} (list) - list column with all plugin data
}}
}
\value{
a \link[tibble:tibble-package]{tibble} with generic data from each trial/plugin result on separate lines
}
\description{
Export all generic data from single session
}
