% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_survey_html_form}
\alias{parse_survey_html_form}
\title{survey-html-form parser}
\usage{
parse_survey_html_form(d)
}
\arguments{
\item{d}{List with unprocessed trial data}
}
\value{
Single row tibble with results. Check \href{https://www.jspsych.org/7.3/plugins/survey-html-form}{jsPsych documentation}
for the list of available variables.
}
\description{
https://www.jspsych.org/7.3/plugins/survey-html-form
}
\examples{
\dontrun{
fn <- demo_file("jspsych-survey-html-form.json")
d  <- read_jspsych(fn)
trials <-
  d \%>\%
  filter(trial_type == trial_types$survey_html_form) \%>\%
  select(record, trial_index, raw) \%>\%
  process_records(.using = parse_survey_html_form) \%>\%
  unnest(processed)
trials \%>\%
  filter(trial_index == 0) \%>\%
  select(-raw) \%>\%
  mutate(bind_rows(response))
}
}
