% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_survey_likert}
\alias{parse_survey_likert}
\title{survey-likert parser}
\usage{
parse_survey_likert(d)
}
\arguments{
\item{d}{List with unprocessed trial data}
}
\value{
Single row tibble with results. Check \href{https://www.jspsych.org/7.3/plugins/survey-likert}{jsPsych documentation}
for the list of available variables.
}
\description{
https://www.jspsych.org/7.3/plugins/survey-likert
}
\examples{
\dontrun{
fn <- demo_file("jspsych-survey-likert.json")
d  <- read_jspsych(fn)
trials <-
  d \%>\%
  filter(trial_type == trial_types$survey_likert) \%>\%
  select(record, trial_index, raw) \%>\%
  process_records(.using = parse_survey_likert) \%>\%
  unnest(processed)
# individual questions/responses
trials \%>\%
  filter(trial_index == 1) \%>\%
  unnest(response)
}
}
