% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_survey_multi_select}
\alias{parse_survey_multi_select}
\title{survey-multi-select parser}
\usage{
parse_survey_multi_select(d)
}
\arguments{
\item{d}{List with unprocessed trial data}
}
\value{
Single row tibble with results. Check \href{https://www.jspsych.org/7.3/plugins/survey-multi-select}{jsPsych documentation}
for the list of available variables.
}
\description{
https://www.jspsych.org/7.3/plugins/survey-multi-select
}
\examples{
\dontrun{
fn <- demo_file("jspsych-survey-multi-select.json")
d  <- read_jspsych(fn)
trials <-
  d \%>\%
  filter(trial_type == trial_types$survey_multi_select) \%>\%
  select(record, trial_index, raw) \%>\%
  process_records(.using = parse_survey_multi_select) \%>\%
  unnest(processed)
trials \%>\%
  select(-raw, -question_order) \%>\%
  filter(trial_index == 0) \%>\%
  unnest(response)
}
}
