% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_survey_text}
\alias{parse_survey_text}
\title{survey-text parser}
\usage{
parse_survey_text(d)
}
\arguments{
\item{d}{List with unprocessed trial data}
}
\value{
Single row tibble with results. Check \href{https://www.jspsych.org/7.3/plugins/survey-text}{jsPsych documentation}
for the list of available variables.
}
\description{
https://www.jspsych.org/7.3/plugins/survey-text
}
\examples{
\dontrun{
fn <- demo_file("jspsych-survey-text.json")
d  <- read_jspsych(fn)
trials <-
  d \%>\%
  filter(trial_type == trial_types$survey_text) \%>\%
  select(record, trial_index, raw) \%>\%
  process_records(.using = parse_survey_text) \%>\%
  unnest(processed)
trials \%>\%
  filter(trial_index == 1) \%>\%
  unnest(response)
}
}
