% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_visual_search_circle}
\alias{parse_visual_search_circle}
\title{visual-search-circle parser}
\usage{
parse_visual_search_circle(d)
}
\arguments{
\item{d}{List with unprocessed trial data}
}
\value{
Single row tibble with results. Check \href{https://www.jspsych.org/7.3/plugins/visual-search-circle}{jsPsych documentation}
for the list of available variables.
}
\description{
https://www.jspsych.org/7.3/plugins/visual-search-circle
}
\examples{
\dontrun{
fn <- demo_file("jspsych-visual-search-circle.json")
d  <- read_jspsych(fn)
trials <-
  d \%>\%
  filter(trial_type == trial_types$visual_search_circle) \%>\%
  select(record, trial_index, raw) \%>\%
  process_records(.using = parse_visual_search_circle) \%>\%
  unnest(processed)
trials \%>\%
  select(record, trial_index, locations) \%>\%
  unnest(locations)
}
}
