% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_jspsych.R
\name{read_jspsych}
\alias{read_jspsych}
\title{Reads jsPsych data into tibble}
\usage{
read_jspsych(filepath, single = T)
}
\arguments{
\item{filepath}{File to read}

\item{single}{Indicates whether the file contains a single record
or multiple records (each on separate line)}
}
\value{
a \link[tibble:tibble-package]{tibble} with each trial/plugin result as a separate line
}
\description{
By default, the code expects a file with a single JSON structure
with white spaces and line delimiters \code{(single = T)}.
The alternative behaviour \code{(single = F)} expects a format with multiple
JSON records, each in a single line.
Results stored in JATOS server follow this structure.
}
