% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{single_value}
\alias{single_value}
\title{Wrapper for possible null objects}
\usage{
single_value(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
The same object or NA if the object is null.
}
\description{
Wrapper for possible null objects
}
\examples{
single_value(1)
single_value(NA)
single_value(NULL)
single_value(1:3)
}
