% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_list.R
\docType{data}
\name{trial_types}
\alias{trial_types}
\title{List of trial types}
\format{
An object of class \code{list} of length 45.
}
\source{
\url{https://www.jspsych.org/7.0/plugins/list-of-plugins/}
}
\usage{
trial_types
}
\description{
List of trial types or plugins from jsPsych. Instead of specifying
a correct name for a plugin, you can start typing \code{trial_types} and
rely on auto-complete. The names in the list are identical to the
original plugin names, only minus signs are replaced with underscores.
}
\examples{
trial_types$html_button_response
}
\keyword{datasets}
