% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{v_integer}
\alias{v_integer}
\alias{v_character}
\alias{v_real}
\alias{v_logical}
\title{Typed wrappers for possible null objects}
\usage{
v_integer(x)

v_character(x)

v_real(x)

v_logical(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
The same object or NA if the object is null.
}
\description{
\code{v_integer()} converts to integers

\code{v_character()} converts to character

\code{v_real()} converts to real numbers

\code{v_logical()} converts to logical values
}
