% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_call_function}
\alias{parse_call_function}
\title{call-function parser}
\usage{
parse_call_function(d)
}
\arguments{
\item{d}{List with unprocessed trial data}
}
\value{
Single row tibble with results. Check \href{https://www.jspsych.org/7.3/plugins/call-function}{jsPsych documentation}
for the list of available variables.
}
\description{
https://www.jspsych.org/7.3/plugins/call-function
}
\examples{
\dontrun{
fn <- demo_file("jspsych-call-function.json")
d  <- read_jspsych(fn)
d \%>\%
  filter(trial_type == trial_types$call_function) \%>\%
  select(record, trial_index, raw) \%>\%
  process_records(.using = parse_call_function) \%>\%
  unnest(processed)
}
}
