% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_instructions}
\alias{parse_instructions}
\title{jspsych-instructions parse}
\usage{
parse_instructions(d)
}
\arguments{
\item{d}{List with unprocessed trial data

view_history    array    An array containing the order of pages the subject viewed (including when the subject returned to previous pages) and the time spent viewing each page. Each object in the array represents a single page view, and contains keys called page_index (the page number, starting with 0) and viewing_time (duration of the page view). This will be encoded as a JSON string when data is saved using the .json() or .csv() functions.
rt    numeric    The response time in milliseconds for the subject to view all of the pages.}
}
\value{
Single row tibble with results
}
\description{
https://www.jspsych.org/plugins/jspsych-instructions/
}
