% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_survey_html_form}
\alias{parse_survey_html_form}
\title{jspsych-survey-html-form parser}
\usage{
parse_survey_html_form(d)
}
\arguments{
\item{d}{List with unprocessed trial data

response    object    An object containing the response for each input. The object will have a separate key (variable) for the response to each input, with each variable being named after its corresponding input element. Each response is a string containing whatever the subject answered for this particular input. This will be encoded as a JSON string when data is saved using the .json() or .csv() functions.
rt    numeric    The response time in milliseconds for the subject to make a response.}
}
\value{
Single row tibble with results
}
\description{
https://www.jspsych.org/plugins/jspsych-survey-html-form/
}
\examples{
fn <- system.file("testdata", "jspsych-survey-html-form.json", package = "jspsychread")
d  <- read_jspsych(fn)
d$raw \%>\%
  purrr::map(~ parse_survey_html_form(.x)) \%>\%
  dplyr::bind_rows()
}
