% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_records.R
\name{process_records}
\alias{process_records}
\title{Process raw jsPsych data}
\usage{
process_records(x, .using = NULL)
}
\arguments{
\item{x}{Dataset with jsPsych records}

\item{.using}{Parser function to use on each record. Use \code{NULL} for auto-detection.}
}
\value{
Original dataset with new column \code{processed} with nested results.
}
\description{
Process raw jsPsych data
}
\examples{
\dontrun{
fn <- demo_file("jspsych-html-button-response.json")
d  <- read_jspsych(fn)
d \%>\%
  filter(trial_type == trial_types$html_button_response) \%>\%
  process_records(.using = parse_html_button_response) \%>\%
  unnest(processed)
}
}
